Restructure directory and file names for documents (#243)
Fixes:
[bluewave-labs/datahall#221](https://github.com/bluewave-labs/datahall/issues/221)

This PR renames the `/links` directory to `/documentAccess` for file
access components, aligning with the requested changes. Additionally, it
updates file names accordingly.

The RowBox component, originally designed to conditionally access
elements in the visitor details form page, is only used in the
VisitorInfoModal component.
And elements are accessed conditional so I had to use `Styled` component
to make it clear and cleaner approach.

Thanks

Co-authored-by: Mahid Ahmad <mahid797@hotmail.com>